requirements.txt
flask~=2.3.2
python-dotenv~=1.0.0
openai~=0.27.8
langchain~=0.0.253
google-cloud-aiplatform~=1.29.0
gunicorn==20.1.0